+++ test/test_helper.exs
@@ -1,7 +1,6 @@
 ExUnit.start(exclude: [:integration])
 Ecto.Adapters.SQL.Sandbox.mode(HexdocsMcp.Repo, :manual)
 
-# Mock definitions
 Mox.defmock(HexdocsMcp.MockOllama, for: HexdocsMcp.Behaviours.Ollama)
 Mox.defmock(HexdocsMcp.MockEmbeddings, for: HexdocsMcp.Behaviours.Embeddings)
 Mox.defmock(HexdocsMcp.MockFetchDocs, for: HexdocsMcp.Behaviours.CLI.FetchDocs)
@@ -11,7 +10,6 @@ Mox.defmock(HexdocsMcp.MockMixDeps, for: HexdocsMcp.Behaviours.MixDeps)
 Mox.defmock(HexdocsMcp.MockHexSearch, for: HexdocsMcp.Behaviours.HexSearch)
 Mox.defmock(HexdocsMcp.MockFulltextSearch, for: HexdocsMcp.Behaviours.FulltextSearch)
 
-# Set default mocks for testing
 Application.put_env(:hexdocs_mcp, :ollama_client, HexdocsMcp.MockOllama)
 Application.put_env(:hexdocs_mcp, :embeddings_module, HexdocsMcp.Embeddings)
 Application.put_env(:hexdocs_mcp, :cli_module, HexdocsMcp.CLI)